requirements.txt
mlx-whisper>=0.5.0
pyannote.audio>=3.1
torch>=2.5.0
torchaudio>=2.5.0
huggingface-hub>=0.27.0
gradio>=5.0.0
python-dotenv>=1.0.0
